+++ packages/importer/src/index.ts
@@ -1,5 +1,5 @@
 export { guessColumnTypes } from './columns';
-export { DEFAULT_CSV_DELIMITERS_TO_GUESS, parseJson } from './parser';
+export { DEFAULT_CSV_DELIMITERS_TO_GUESS } from './parsers/csvParser';
 export * from './plugin';
 export type { SyncCsvResults } from './types';
 export { generateRandomData } from './random-data';
