--- packages/importer/src/utils/dates.ts
@@ -1,10 +1,0 @@
-import { Temporal } from '@js-temporal/polyfill';
-
-export function isValidDate(value: any): boolean {
-  try {
-    Temporal.PlainDateTime.from(value);
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
